backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.89.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.13.0.92
packaging==26.0
pillow==12.2.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.33.1
strawberry-graphql==0.314.3
SQLAlchemy==2.0.49
PyYAML==6.0.3
regex==2026.4.4
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
ultralytics==8.3.85
openai==1.52.2
python-multipart==0.0.12
